sync/requirements.txt
wget==3.2
ruamel.yaml==0.16.12
google-cloud-storage==1.23.0
Jinja2==2.11.1
google-auth==1.14.0
absl-py==0.9.0
urlopen==1.0.0
markdown==3.1.1
lxml==4.5.2
coverage==5.3
flake8==3.8.3
